requirements.txt
google-generativeai
duckduckgo-search
pillow
pandas
python-dotenv
imagehash
google-api-python-client
